test: split oneclick tests on inscription and transaction
